+++ sope-xml/SaxObjC/SaxHandlerBase.m
@@ -38,9 +38,9 @@
 - (void)endElement:(NSString *)_tagName {
 }
 
-- (void)characters:(unichar *)_chars length:(int)_len {
+- (void)characters:(unichar *)_chars length:(NSUInteger)_len {
 }
-- (void)ignorableWhitespace:(unichar *)_chars length:(int)_len {
+- (void)ignorableWhitespace:(unichar *)_chars length:(NSUInteger)_len {
 }
 
 - (void)processingInstruction:(NSString *)_pi data:(NSString *)_data {
