+++ sope-xml/SaxObjC/SaxXMLFilter.m
@@ -155,10 +155,10 @@
   [self->contentHandler endElement:_localName namespace:_ns rawName:_rawName];
 }
 
-- (void)characters:(unichar *)_chars length:(int)_len {
+- (void)characters:(unichar *)_chars length:(NSUInteger)_len {
   [self->contentHandler characters:_chars length:_len];
 }
-- (void)ignorableWhitespace:(unichar *)_chars length:(int)_len {
+- (void)ignorableWhitespace:(unichar *)_chars length:(NSUInteger)_len {
   [self->contentHandler ignorableWhitespace:_chars length:_len];
 }
 - (void)processingInstruction:(NSString *)_pi data:(NSString *)_data {
